Moved environment constants to constants.py for consistency
